70: refactored custom plugin project factory
Task-Url: https://github.com/LorenzoBettini/javamm/issues/issue/70
